+++ src/inspect_ai/_eval/run.py
@@ -215,6 +215,12 @@ async def eval_run(
                 else:
                     task.continue_on_fail = task_eval_config.continue_on_fail
 
+                # score_on_error
+                if task_eval_config.score_on_error is None:
+                    task_eval_config.score_on_error = task.score_on_error
+                else:
+                    task.score_on_error = task_eval_config.score_on_error
+
                 # merge eval-level and task-level tags
                 merged_tags = list(set(tags or []) | set(task.tags or [])) or None
 
@@ -268,6 +274,8 @@ async def eval_run(
                         debug_errors=debug_errors,
                         sample_source=resolved_task.sample_source,
                         kwargs=kwargs,
+                        initial_model_usage=resolved_task.initial_model_usage,
+                        initial_role_usage=resolved_task.initial_role_usage,
                     )
                 )